refactor: StatusCode enum의 code와 메시지를 분리
