+++ src/main/java/http/header/MimeType.java
@@ -1,4 +1,4 @@
-package http;
+package http.header;
 
 import java.util.Set;
