+++ src/main/java/http/statusline/HttpMethod.java
@@ -1,4 +1,4 @@
-package http;
+package http.statusline;
 
 public enum HttpMethod {
 	GET, POST, PUT, DELETE, CONNECT, HEAD, OPTIONS, TRACE, NOT_ANNOUNCED;
