+++ src/main/java/http/statusline/HttpVersion.java
@@ -1,4 +1,4 @@
-package http;
+package http.statusline;
 
 public enum HttpVersion {
 	HTTP_1_1("HTTP/1.1"),
