+++ src/main/java/webserver/RequestHandler.java
@@ -15,10 +15,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import annotations.AnnotationMap;
-import http.HttpMethod;
+import http.statusline.HttpMethod;
 import http.HttpRequest;
 import http.HttpResponse;
-import http.StatusCode;
+import http.statusline.StatusCode;
 
 public class RequestHandler implements Runnable {
 	private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
